Convert Projects section to typescript
